Warning patch for Win32 - casted off_t to intmax_t for printing.
CVS patchset: 6979
CVS date: 2004/09/17 19:21:33

+++ src/demuxers/demux_avi.c
@@ -19,7 +19,7 @@
  */
 
 /*
- * $Id: demux_avi.c,v 1.207 2004/07/04 21:27:23 mroi Exp $
+ * $Id: demux_avi.c,v 1.208 2004/09/17 19:21:33 valtri Exp $
  *
  * demultiplexer for avi streams
  *
@@ -639,9 +639,7 @@ static int idx_grow(demux_avi_t *this, int (*stopper)(demux_avi_t *, void *),
       }
     }
     if (!valid_chunk) {
-      xine_log (this->stream->xine, XINE_LOG_MSG,
-                _("demux_avi: invalid avi chunk \"%c%c%c%c\" at pos %lld\n"),
-                data[0], data[1], data[2], data[3], chunk_pos);
+      xine_log(this->stream->xine, XINE_LOG_MSG, _("demux_avi: invalid avi chunk \"%c%c%c%c\" at pos %" PRIiMAX "\n"), data[0], data[1], data[2], data[3], (intmax_t)chunk_pos);
     }
     chunk_pos = this->input->seek(this->input, this->idx_grow.nexttagoffset, SEEK_SET);
     if (chunk_pos != this->idx_grow.nexttagoffset) {
@@ -784,8 +782,7 @@ static avi_t *AVI_init(demux_avi_t *this) {
     this->idx_grow.nexttagoffset = this->input->get_current_pos(this->input);
 
     if (this->input->read(this->input, data,8) != 8 ) {
-      xprintf(this->stream->xine, XINE_VERBOSITY_LOG, "failed to read 8 bytes at pos %lld\n",
-              this->idx_grow.nexttagoffset);
+      xprintf(this->stream->xine, XINE_VERBOSITY_LOG, "failed to read 8 bytes at pos %" PRIiMAX "\n", (intmax_t)this->idx_grow.nexttagoffset);
       break; /* We assume it's EOF */
     }
 
@@ -849,9 +846,7 @@ static avi_t *AVI_init(demux_avi_t *this) {
       }
     }
     if (next_chunk != this->input->seek(this->input, next_chunk, SEEK_SET)) {
-      xine_log (this->stream->xine, XINE_LOG_MSG,
-                _("demux_avi: failed to seek to the next chunk (pos %lld)\n"),
-                next_chunk);
+      xine_log (this->stream->xine, XINE_LOG_MSG, _("demux_avi: failed to seek to the next chunk (pos %" PRIiMAX ")\n"), (intmax_t)next_chunk);
       break;  /* probably slow seek */
     }
   }
@@ -1691,9 +1686,7 @@ static int get_chunk_header(demux_avi_t *this, uint32_t *len, int *audio_stream)
         return AVI_HEADER_AUDIO;
       }
     }
-    xine_log (this->stream->xine, XINE_LOG_MSG,
-              _("demux_avi: invalid avi chunk \"%c%c%c%c\" at pos %lld\n"),
-              data[0], data[1], data[2], data[3], this->input->get_current_pos(this->input));
+    xine_log (this->stream->xine, XINE_LOG_MSG, _("demux_avi: invalid avi chunk \"%c%c%c%c\" at pos %" PRIiMAX "\n"), data[0], data[1], data[2], data[3], (intmax_t)this->input->get_current_pos(this->input));
     return AVI_HEADER_UNKNOWN;
   }
   /* unreachable code */
@@ -2080,8 +2073,9 @@ static int demux_avi_seek_internal (demux_avi_t *this) {
    * seek to start pos / time
    */
 
-  xprintf (this->stream->xine, XINE_VERBOSITY_DEBUG,
-	   "start pos is %lld, start time is %d\n",start_pos, start_time);
+  xprintf(this->stream->xine, XINE_VERBOSITY_DEBUG,
+          "start pos is %" PRIiMAX ", start time is %d\n", (intmax_t)start_pos,
+          start_time);
 
   /* Seek video.  We do a single idx_grow at the beginning rather than
    * incrementally growing the index in a loop, so that if the index
