+++ src/xine-engine/input_rip.c
@@ -29,7 +29,7 @@
  * - it's possible speeder saving streams in the xine without playing:
  *     xine stream_mrl#save:file.raw\;noaudio\;novideo
  *
- * $Id: input_rip.c,v 1.22 2004/05/09 17:42:23 valtri Exp $
+ * $Id: input_rip.c,v 1.23 2004/09/17 19:21:46 valtri Exp $
  */
 
 /* TODO:
@@ -126,8 +126,7 @@ static off_t rip_plugin_read(input_plugin_t *this_gen, char *buf, off_t len) {
   if (nread_file) {
     lprintf(" => read %lld bytes from file\n", nread_file);
     if (fread(&buf[npreview], nread_file, 1, this->file) != 1) {
-      xine_log(this->stream->xine, XINE_LOG_MSG,
-	       _("input_rip: reading of saved data failed: %s\n"), strerror(errno));
+      xine_log(this->stream->xine, XINE_LOG_MSG, _("input_rip: reading of saved data failed: %s\n"), strerror(errno));
       return -1;
     }
   }
@@ -142,7 +141,7 @@ static off_t rip_plugin_read(input_plugin_t *this_gen, char *buf, off_t len) {
 
     if (retlen < 0) {
       xine_log(this->stream->xine, XINE_LOG_MSG, 
-	       _("input_rip: reading by input plugin failed\n"));
+               _("input_rip: reading by input plugin failed\n"));
       return -1;
     }
 
@@ -150,9 +149,7 @@ static off_t rip_plugin_read(input_plugin_t *this_gen, char *buf, off_t len) {
     if (retlen > nread_orig) {
       nwrite = retlen - nread_orig;
       if (fwrite(buf + this->savepos - this->curpos, nwrite, 1, this->file) != 1) {
-        xine_log(this->stream->xine, XINE_LOG_MSG, 
-		 _("input_rip: error writing to file %lld bytes: %s\n"), 
-          retlen - nread_orig, strerror(errno));
+        xine_log(this->stream->xine, XINE_LOG_MSG, _("input_rip: error writing to file %" PRIiMAX " bytes: %s\n"), (intmax_t)(retlen - nread_orig), strerror(errno));
         return -1;
       }
       this->savepos += nwrite;
@@ -173,7 +170,7 @@ static int rip_plugin_open(input_plugin_t *this_gen) {
   rip_input_plugin_t *this = (rip_input_plugin_t *)this_gen;
 
   xine_log(this->stream->xine, XINE_LOG_MSG,
-	   _("input_rip: open() function should never be called\n"));
+           _("input_rip: open() function should never be called\n"));
   return 0;
 }
 
@@ -249,7 +246,8 @@ static buf_element_t *rip_plugin_read_block(input_plugin_t *this_gen, fifo_buffe
       lprintf(" => read %lld bytes from the file (block)\n", nread_file);
       if (fread(&buf->content[npreview], nread_file, 1, this->file) != 1) {
         xine_log(this->stream->xine, XINE_LOG_MSG,
-		 _("input_rip: reading of saved data failed: %s\n"), strerror(errno));
+                 _("input_rip: reading of saved data failed: %s\n"),
+                 strerror(errno));
         return NULL;
       }
     }
@@ -280,8 +278,8 @@ static buf_element_t *rip_plugin_read_block(input_plugin_t *this_gen, fifo_buffe
       nwrite = retlen - nread_orig;
       if (fwrite(buf->content + this->savepos - this->curpos, nwrite, 1, this->file) != 1) {
         xine_log(this->stream->xine, XINE_LOG_MSG, 
-		 _("input_rip: error writing to file %lld bytes: %s\n"), 
-          retlen - nread_orig, strerror(errno));
+                 _("input_rip: error writing to file %" PRIiMAX " bytes: %s\n"),
+                 (intmax_t)(retlen - nread_orig), strerror(errno));
         return NULL;
       }
       this->savepos += nwrite;
@@ -385,8 +383,9 @@ static off_t rip_plugin_seek(input_plugin_t *this_gen, off_t offset, int origin)
     } else {
       if ((pos = rip_seek_original(this, this->savepos)) == -1) return -1;
       if (pos > this->savepos)
-        xine_log(this->stream->xine, XINE_LOG_MSG, _("input_rip: %lld bytes dropped\n"), 
-		 pos - this->savepos);
+        xine_log(this->stream->xine, XINE_LOG_MSG,
+                 _("input_rip: %" PRIiMAX " bytes dropped\n"), 
+                 (intmax_t)(pos - this->savepos));
     }
   }
 
@@ -631,7 +630,8 @@ input_plugin_t *_x_rip_plugin_get_instance (xine_stream_t *stream, const char *f
   if (this->preview && this->preview_size) {
     if (fwrite(this->preview, this->preview_size, 1, this->file) != 1) {
       xine_log(this->stream->xine, XINE_LOG_MSG, 
-	       _("input_rip: error writing to file %lld bytes: %s\n"), this->preview_size, strerror(errno));
+               _("input_rip: error writing to file %" PRIiMAX " bytes: %s\n"),
+               (intmax_t)(this->preview_size), strerror(errno));
       fclose(this->file);
       free(this);
       return NULL;
